refactor: move tags function into utils/registry
