+++ src/lib/utils/api/index.ts
@@ -1,3 +1,4 @@
 export * from './auth';
 export * from './delete';
 export * from './health';
+export * from './registry';
